the game is now in a package format
this helps with importing sibling scripts without circular import errors
